Updated latest to current

+++ CoreDataMigration-Example/Extensions/FileManager/FileManager+ApplicationSupport.swift
@@ -15,8 +15,9 @@ extension FileManager {
     // MARK: - ApplicationSupport
     
     static func clearApplicationSupportDirectoryContents() {
-        let applicationSupportURL = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first!
-        let applicationSupportDirectoryContents = try! FileManager.default.contentsOfDirectory(atPath: applicationSupportURL.path)
+        guard let applicationSupportURL = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask).first, let applicationSupportDirectoryContents = try? FileManager.default.contentsOfDirectory(atPath: applicationSupportURL.path) else {
+            return
+        }
         applicationSupportDirectoryContents.forEach {
             let fileURL = URL(fileURLWithPath: applicationSupportURL.path, isDirectory: true).appendingPathComponent($0)
             try? FileManager.default.removeItem(atPath: fileURL.path)
